Add simple integration tests for all output formats
Summary
--

I spun this off from #19133 to be sure to get an accurate baseline before
modifying any of the formats. I picked the code snippet to include a lint
diagnostic with a fix, one without a fix, and one syntax error. I'm happy to
expand it if there are any other kinds we want to test.

Test Plan
--

New CLI tests