Move JUnit rendering to `ruff_db`
Summary
--

This PR moves the JUnit output format to the new rendering infrastructure. As I
mention in a TODO in the code, there's some code that will be shared with the
`grouped` output format. Hopefully I'll have that PR up too by the time this one
is reviewed.

Test Plan
--

Existing tests moved to `ruff_db`