chore(deps): upgrade mkdocs-github-admonitions-plugin to v0.1.1
Renovate doesn't upgrade sha-pinned dependencies.

The pin was introduced due to the bug fix not being available in a
released version in PR #18163

Now the bug fix has been released so the pin can be removed.

In the future, in such cases, it might be a good idea to add a comment
and/or set a reminder to look back and consider removing the pin
as renovate will not handle sha pins.

As renovate handles almost all dep upgrades, this provides a false sense
of "security" that deps are always kept up to date automatically.

Related to issue #19369

Thanks @ntBre for the pointer!